Front-end Function Development
FineTech is the entrepreneurship project that I initiated in Sept 2023. Now as we have registered as a legal company in Hong Kong, we are developing its functions to solve real-world healthcare problems with AI technologies (especially AI fine-tuning and prompt engineering).

I devoted this part of the project to FineTech's front-end function development. Here are part of the codes.